+++ Cargo.toml
@@ -9,16 +9,5 @@ edition = "2021"
 logos = "0.13.0"
 rustc-hash = "1.1.0"
 
-[dependencies.nom]
-version = "7.1.3"
-default-features = false
-features = ["alloc"]
-
 [[example]]
 name = "demo"
-
-[[example]]
-name = "experiment"
-
-[[example]]
-name = "experiment2"
